ALS-4461: Remove IOExceptions thrown from FBBIS
