+++ etl/src/main/java/edu/harvard/hms/dbmi/avillach/hpds/etl/genotype/MultialleleCounter.java
@@ -4,7 +4,6 @@ import java.io.*;
 import java.util.ArrayList;
 import java.util.TreeSet;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.zip.GZIPInputStream;
 
 import edu.harvard.hms.dbmi.avillach.hpds.data.genotype.VariantMasks;
 import edu.harvard.hms.dbmi.avillach.hpds.data.genotype.VariantSpec;
@@ -22,26 +21,22 @@ public class MultialleleCounter {
             currentChromosome.keys().parallelStream().forEach((offsetBucket) -> {
                 System.out.println("Starting bucket : " + offsetBucket);
                 ConcurrentHashMap<String, VariantMasks> maskMap;
-                try {
-                    maskMap = currentChromosome.get(offsetBucket);
+                maskMap = currentChromosome.get(offsetBucket);
 
-                    TreeSet<VariantSpec> variantsSortedByOffset = new TreeSet<VariantSpec>();
-                    for (String variant : maskMap.keySet()) {
-                        variantsSortedByOffset.add(new VariantSpec(variant));
-                    }
-                    ArrayList<VariantSpec> variantsSortedByOffsetList = new ArrayList(variantsSortedByOffset);
-                    for (int y = 1; y < variantsSortedByOffsetList.size(); y++) {
-                        if (variantsSortedByOffsetList.get(y).metadata.offset.equals(variantsSortedByOffsetList.get(y - 1).metadata.offset)) {
-                            try {
-                                System.out.println("Matching offsets : " + variantsSortedByOffsetList.get(y - 1).specNotation() + " : " + variantsSortedByOffsetList.get(y).specNotation() + ":" + maskMap.get(variantsSortedByOffsetList.get(y - 1).specNotation()).heterozygousMask.toString(2) + ":" + ":" + maskMap.get(variantsSortedByOffsetList.get(y).specNotation()).heterozygousMask.toString(2));
-                            } catch (NullPointerException e) {
-                                System.out.println("Matching offsets : " + variantsSortedByOffsetList.get(y - 1).specNotation() + " : " + variantsSortedByOffsetList.get(y).specNotation());
-                            }
+                TreeSet<VariantSpec> variantsSortedByOffset = new TreeSet<>();
+                for (String variant : maskMap.keySet()) {
+                    variantsSortedByOffset.add(new VariantSpec(variant));
+                }
+                ArrayList<VariantSpec> variantsSortedByOffsetList = new ArrayList<>(variantsSortedByOffset);
+                for (int y = 1; y < variantsSortedByOffsetList.size(); y++) {
+                    if (variantsSortedByOffsetList.get(y).metadata.offset.equals(variantsSortedByOffsetList.get(y - 1).metadata.offset)) {
+                        try {
+                            System.out.println("Matching offsets : " + variantsSortedByOffsetList.get(y - 1).specNotation() + " : " + variantsSortedByOffsetList.get(y).specNotation() + ":" + maskMap.get(variantsSortedByOffsetList.get(y - 1).specNotation()).heterozygousMask.toString(2) + ":" + ":" + maskMap.get(variantsSortedByOffsetList.get(y).specNotation()).heterozygousMask.toString(2));
+                        } catch (NullPointerException e) {
+                            System.out.println("Matching offsets : " + variantsSortedByOffsetList.get(y - 1).specNotation() + " : " + variantsSortedByOffsetList.get(y).specNotation());
                         }
                     }
-                } catch (IOException e) {
-					throw new UncheckedIOException(e);
-				}
+                }
                 System.out.println("Completed bucket : " + offsetBucket);
             });
         }
